Make shapeFactor a double instead of a uint8_8. (#88)
In C++, a unit8_t is usually handled as an unsigned char. That means
that extracting such value from a stream will give unexpected results:
if the stream contains "1", the *character* '1' will be extracted,
which has numberical value 49. When converting to a double (as happened
in the code), 49 is used as value instead of 1 for the shapeParameter.
This popped up while working on and testing the prescribed(Eedf) type.

As an example, compile and run the following C++ file:

    #include <cstdint>
    #include <iostream>
    #include <sstream>

    int main()
    {
        std::stringstream ss; ss << "1";
        uint8_t shapeParameter;
        ss >> shapeParameter;
        const double d = shapeParameter;
        std::cout << shapeParameter << std::endl;
        std::cout << d << std::endl;
        return 0;
    }

This will print 1 49, the *character* 1 and its ASCII value 49.

+++ source/ElectronKinetics.cpp
@@ -1361,7 +1361,7 @@ ElectronKineticsPrescribed::ElectronKineticsPrescribed(const std::filesystem::pa
 
 ElectronKineticsPrescribed::ElectronKineticsPrescribed(const std::filesystem::path &basePath,const json_type &cnf, WorkingConditions *workingConditions)
 : ElectronKinetics(basePath,cnf,workingConditions),
-  shapeParameter(cnf.at("shapeParameter").get<unsigned>())
+  shapeParameter(cnf.at("shapeParameter").get<double>())
 {
     if (getIonizationOperatorType(cnf.at("ionizationOperatorType")) != IonizationOperatorType::conservative)
     {
